+++ client/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod client;
 pub mod encryption;
 pub mod error;
+pub mod logic;
 pub mod net;
 pub mod storage;
 pub mod time;
