+++ client/src/logic/mod.rs
@@ -1,0 +1,6 @@
+mod logic;
+
+pub use logic::{
+    fetch_prekeys, process_messages, provision_device, publish_prekeys, register_account,
+    send_message,
+};
