+++ e2e/tests/message.rs
@@ -5,19 +5,14 @@ use std::time::Duration;
 use libsignal_protocol::IdentityKeyPair;
 use rstest::rstest;
 use sam_client::{
+    client::SqliteClientType,
     encryption::envelope::DecryptedEnvelope,
     net::{
         http_client::HttpClientConfig,
-        protocol::{
-            client::ProtocolClient, traits::SamProtocolClient, WebSocketProtocolClientConfig,
-        },
+        protocol::WebSocketProtocolClientConfig,
         tls::{create_tls_config, MutualTlsConfig},
-        ApiClient, HttpClient,
     },
-    storage::{
-        sqlite::{SqliteStoreConfig, SqliteStoreType},
-        StoreType,
-    },
+    storage::sqlite::SqliteStoreConfig,
     Client, ClientError,
 };
 use sam_common::{api::LinkDeviceToken, AccountId};
@@ -29,11 +24,7 @@ use crate::utils::server::TestServer;
 
 const TIMEOUT_SECS: u64 = 120;
 
-async fn client(
-    address: &str,
-    username: &str,
-    device_name: &str,
-) -> Client<SqliteStoreType, HttpClient, ProtocolClient> {
+async fn client(address: &str, username: &str, device_name: &str) -> Client<SqliteClientType> {
     Client::from_registration()
         .username(username)
         .device_name(device_name)
@@ -51,7 +42,7 @@ async fn tls_client(
     username: &str,
     device_name: &str,
     mutual_config: Option<MutualTlsConfig>,
-) -> Client<SqliteStoreType, HttpClient, ProtocolClient> {
+) -> Client<SqliteClientType> {
     let client_config =
         create_tls_config("./cert/rootCA.crt", mutual_config).expect("Can create client config");
     Client::from_registration()
@@ -77,7 +68,7 @@ async fn client_device(
     device_name: &str,
     id_pair: IdentityKeyPair,
     token: LinkDeviceToken,
-) -> Client<SqliteStoreType, HttpClient, ProtocolClient> {
+) -> Client<SqliteClientType> {
     Client::from_provisioning()
         .store_config(SqliteStoreConfig::in_memory().await)
         .api_client_config(HttpClientConfig::new(address.to_string()))
@@ -418,8 +409,8 @@ enum Message<'a> {
 }
 
 async fn send(
-    sender: &mut Client<impl StoreType, impl ApiClient, impl SamProtocolClient>,
-    receiver: &mut Client<impl StoreType, impl ApiClient, impl SamProtocolClient>,
+    sender: &mut Client<SqliteClientType>,
+    receiver: &mut Client<SqliteClientType>,
     id: AccountId,
     subscriber: &mut Receiver<DecryptedEnvelope>,
     msg: &str,
@@ -491,7 +482,7 @@ async fn sqlite_stores_alice_send_to_bob() {
         let temp = NamedTempFile::new().expect("Can create tempfile");
         let path = format!("sqlite://{}?mode=rwc", temp.path().to_string_lossy());
 
-        let mut alice = Client::from_registration()
+        let mut alice: Client<SqliteClientType> = Client::from_registration()
             .username("Alice")
             .device_name("Alice's Device")
             .store_config(SqliteStoreConfig::new(path.clone()))
@@ -509,7 +500,7 @@ async fn sqlite_stores_alice_send_to_bob() {
             .await
             .expect("can create a store");
 
-        let mut alice = Client::from_store()
+        let mut alice: Client<SqliteClientType> = Client::from_store()
             .store(store)
             .protocol_config(WebSocketProtocolClientConfig::new(address.to_owned()))
             .api_client_config(HttpClientConfig::new(address.to_owned()))
@@ -517,7 +508,7 @@ async fn sqlite_stores_alice_send_to_bob() {
             .await
             .unwrap();
 
-        let bob = Client::from_registration()
+        let bob: Client<SqliteClientType> = Client::from_registration()
             .username("Bob")
             .device_name("Bob's Device")
             .store_config(SqliteStoreConfig::new("sqlite::memory:".to_owned()))
